Fix warnings in the example directory.

+++ example/hx_controller.c
@@ -150,7 +150,8 @@ int main(int argc, char **argv)
     for (i = 0; i < robotInfo.motor_count; ++i)
     {
       // Set the desired position of this motor
-      cmd.ref_pos[i] = 350 * 0.5 * sin(0.05 * 2.0 * M_PI * counter * 0.01);
+      cmd.ref_pos[i] = (float)(350 * 0.5 *
+        sin(0.05 * 2.0 * M_PI * counter * 0.01));
       // We could set a desired maximum velocity
       // cmd.ref_vel[i] = 1.0;
       // cmd.ref_vel_max[i] = 1.0;
@@ -199,7 +200,7 @@ int main(int argc, char **argv)
     // executed by the limb).
     unsigned int sleeptime_us = 10000;
 #ifdef _WIN32
-    Sleep(sleeptime_us / 1e3);
+    Sleep((DWORD)(sleeptime_us / 1e3));
 #else
     usleep(sleeptime_us);
 #endif
